+++ app/src/main/java/com/rick/posadka/model/PosadkaShaft.kt
@@ -15,9 +15,9 @@ data class PosadkaShaft(
     @ColumnInfo(name = "class")
     val name: String,
     @ColumnInfo(name = "highdeviation")
-    val highDeviation: Double,
+    val highDeviation: Float,
     @ColumnInfo(name = "lowdeviation")
-    val lowDeviation: Double
+    val lowDeviation: Float
 
 )
